changed Pagination total to Long data type added setters for total and results

+++ devnull-util/src/main/groovy/org/devnull/util/pagination/SimplePagination.groovy
@@ -27,7 +27,7 @@ class SimplePagination<T extends Serializable> implements Pagination<T>, Seriali
     /**
      * Default = {@value}
      */
-    Integer total = 0
+    Long total = 0
 
     /**
      * Default = {@value}
